rename map.rb to mapdata.rb
