+++ main.rb
@@ -4,7 +4,7 @@ require_relative 'director'
 require_relative 'menu'
 require_relative 'cell'
 require_relative 'mapfactory'
-require_relative 'map'
+require_relative 'mapdata'
 
 Window.caption = "RE:source Defence"
 #Window.width = 800
